Create 5kyu_FaultyOdometr.cs
The odometer has a defect: it proceeds from digit 3 to digit 5 always skipping the digit 4. This defect shows up in all positions (ones, tens, hundreds, etc).
For example, if the odometer displays 15339 and the car travels another mile, the odometer changes to 15350 (instead of 15340).

Writing a function to calculate the real distance, according The number the odometer shows.