Create 6kyu_MinutesToMidnight.cs
Teemo is not really excited about the new year's eve, but he has to celebrate it with his friends anyway. 
He has a really big passion about programming and he want's to be productive till midnight. He want's to know how many minutes he has left to work on his new project.
He doesn't want to look on the clock all the time, so he thought about a function, which returns him the number of minutes.